fix(server): count every stored table against the account cap
`ensure_account_capacity` is the only enforcement point for
`CHUR_SYNC_MAX_ACCOUNT_BYTES`, which docs/sync/SERVER_OPERATOR.md
defines as the maximum reserved bytes for one vault identity. Its sum
named six tables and missed two: `collection_operations` and
`deletion_requests`.

For `collection_operations` that made the check a permanent no-op.
`accept_collection_operation` calls the gate and then inserts into the
one table the gate does not measure, so the running total never grew and
every subsequent call saw the same figure. A collection operation
carries up to a mebibyte of payload, so any enrolled device -- or any
CONTRIBUTE recipient of any shared collection -- could append a megabyte
per request until the operator's disk was full, at which point every
other vault on the deployment starts failing.

`begin_upload` held a second copy of the same query with the same two
omissions, so object uploads did not see the space either. That copy is
why one omission became two, and it is now the shared call. The width
moves to `u64` because that is what `max_account_bytes` and the upload
length already are.

`transport_tokens` stays out on purpose: one digest per enrolled device,
and enrolment is already charged through `operations` and
`membership_records`. `deletion_requests` counts because the retained
receipt belongs to that identity, and no deletion route calls the gate,
so a full account can still delete itself.

+++ rust/crates/chur-sync-server/src/checkpoint.rs
@@ -113,7 +113,7 @@ impl ReferenceServer {
                 "checkpoint head is absent or differs"
             );
         }
-        self.ensure_account_capacity(checkpoint.vault_id(), checkpoint.encode().len())?;
+        self.ensure_account_capacity(checkpoint.vault_id(), checkpoint.encode().len() as u64)?;
         self.db
             .execute(
                 "INSERT INTO checkpoints (
